Fixed restructuring issues
